Bug 1082598 - Part 1: Update Skia to master revision 53c5d5fb795fe04bec050c0583223027c25b839b. r=jrmuizel
UltraBlame original commit: aa380501527094672fd84092655f82e62c504482

+++ gfx/skia/skia/src/core/SkDistanceFieldGen.h
@@ -35,7 +35,7 @@
 
 bool SkGenerateDistanceFieldFromA8Image(unsigned char* distanceField,
                                         const unsigned char* image,
-                                        int w, int h, int rowBytes);
+                                        int w, int h, size_t rowBytes);
 
 
 
@@ -48,7 +48,7 @@ bool SkGenerateDistanceFieldFromA8Image(unsigned char* distanceField,
 
 bool SkGenerateDistanceFieldFromBWImage(unsigned char* distanceField,
                                         const unsigned char* image,
-                                        int w, int h, int rowBytes);
+                                        int w, int h, size_t rowBytes);
